+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "timr",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.17.0",
