+++ src/Components/ConfigurationPanel/index.js
@@ -1,0 +1,2 @@
+// import * from './ConfigurationPanel'
+export { default } from './ConfigurationPanel'
