+++ src/Components/icon/Icon.js
@@ -1,0 +1,7 @@
+const Icon = ({icon}) => {
+    return (
+        <div>icon</div>
+    )
+}
+
+export default Icon
