+++ src/Components/icon/index.js
@@ -1,7 +1,1 @@
-const Icon = ({icon}) => {
-    return (
-        <div>icon</div>
-    )
-}
-
-export default Icon
+export { default } from './Icon'
